reorg(0254): extract corpus predicates to flat _corpus_predicates.py
build_het_core.py was dual-role: the Phase-2 core-set builder, but also
imported by analyze_multilingual.py for its is_global_south / is_non_english
row predicates. That script→script import blocks the wave-1 move.

Relocate GLOBAL_SOUTH_KEYWORDS + is_non_english / is_global_south verbatim
into a neutral flat _corpus_predicates.py (the 0250 pattern) and repoint
analyze_multilingual.py. build_het_core.py re-imports the two predicates it
still uses internally; the keyword set is used only inside is_global_south so
it moves wholesale. Output byte-identical by construction (keyword set and
bodies unchanged; verified on representative rows). build_het_core.py is now a
pure entry point, free to move in wave-1 (0257).

Ticket-ref: tickets/0254-reorg-code-entry-point-vs-library-classi.erg
